drag_python_client/requirements.txt
web3>=6.0.0
eth-account>=0.11.0
pyyaml>=6.0.0

